% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/superlearner_base.R
\name{base_wflow}
\alias{base_wflow}
\alias{base_wflow.default}
\alias{base_wflowsmall}
\alias{print.base_wflow}
\alias{print.base_wflowsmall}
\title{base_wflow}
\usage{
base_wflow(
  dataset,
  outcome,
  rep_specs = 3,
  times = 4,
  rango = c(0.3, 0.6),
  engines = NULL,
  ns = TRUE,
  poly = TRUE,
  add_step = NULL,
  rango_cols = c(0.25, 0.7)
)

\method{base_wflow}{default}(
  dataset,
  outcome,
  rep_specs = 3,
  times = 4,
  rango = c(0.3, 0.6),
  engines = NULL,
  ns = TRUE,
  poly = TRUE,
  add_step = NULL,
  rango_cols = c(0.25, 0.7)
)

base_wflowsmall(
  dataset,
  outcome,
  rep_specs = 3,
  times = 4,
  rango = c(0.7, 0.8),
  engines = NULL,
  add_step = NULL
)

\method{print}{base_wflow}(x)

\method{print}{base_wflowsmall}(x)
}
\value{
base_wflow
}
\description{
base_wflow

base de construccion de modelos

predict sl clas
}
