% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bestnorm.R
\name{bestnorm}
\alias{bestnorm}
\alias{bestnorm_transfor}
\alias{bestnorm_fn}
\alias{bestnorm_vec}
\title{buscar la mejor normalizacion de las variables}
\usage{
bestnorm(.data, ...)

bestnorm_transfor(x)

bestnorm_fn(x)

bestnorm_vec(x)
}
\arguments{
\item{.data}{dataset}

\item{...}{columnas a normalizar}

\item{x}{vector a normalizar}
}
\value{
normalizacion de variables
}
\description{
objeto de la mejor normalizacion: bestnorm

evaluacion de todos los metodos de

respuesta de la funcion de transformacion optima: bestnorm_fn

normalizacion de un vector: bestnorm_vec
}
\examples{
credit <- mgalda::datalearn$credit
set.seed(7)
object <- bestnorm(.data = credit, where(is_dblint))
predict(object = object, new_data = credit)
autoplot(object)

x <- mgalda::datalearn$credit$price
set.seed(7)
object <- bestnorm_transfor(x)
object

x <- mgalda::datalearn$credit$price
set.seed(7)
fbest <- bestnorm_fn(x)
autoplot(fbest)

x <- mgalda::datalearn$credit$price
set.seed(7)
plot(x, bestnorm_vec(x))

x <- mgalda::datalearn$credit$price
set.seed(7)
bestnorm_transfor(x)
x <- mgalda::datalearn$credit$price
set.seed(7)
autoplot(bestnorm_fn(x))
x <- mgalda::datalearn$credit$price
set.seed(7)
bestnorm_vec(x)
}
