% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/casewhen_constructor.R
\name{casewhen_constructor}
\alias{casewhen_constructor}
\alias{variable.names.casewhen_constructor}
\alias{print.casewhen_constructor}
\alias{fn_cw}
\alias{fn_cw.casewhen_constructor}
\alias{.casewhen_constructor}
\alias{plural}
\title{Crear expreciones case_when}
\usage{
casewhen_constructor(..., vars = "x")

\method{variable.names}{casewhen_constructor}(object, ...)

\method{print}{casewhen_constructor}(x, ...)

fn_cw(x, ...)

\method{fn_cw}{casewhen_constructor}(x, ...)

.casewhen_constructor(..., vars, fn = dplyr::case_when)

plural(x, plural = "s")
}
\arguments{
\item{vars}{argumento a evaluar}
}
\description{
Crear expreciones case_when

Crear expreciones case_when
}
\keyword{internal}
