% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/custom_metrics.R
\name{custom_metrics}
\alias{custom_metrics}
\alias{nrmse}
\alias{nrmse.data.frame}
\alias{nrmse_vec}
\alias{rsq_sliding}
\alias{rsq_sliding.data.frame}
\alias{rsq_sliding_vec}
\alias{w_accuracy}
\alias{w_accuracy.data.frame}
\alias{w_accuracy_vec}
\title{custom_metrics}
\usage{
nrmse(data, ...)

\method{nrmse}{data.frame}(data, truth, estimate, na_rm = TRUE, ...)

nrmse_vec(truth, estimate, na_rm = TRUE, ...)

rsq_sliding(data, ...)

\method{rsq_sliding}{data.frame}(data, truth, estimate, na_rm = TRUE, ...)

rsq_sliding_vec(truth, estimate, na_rm = TRUE, ...)

w_accuracy(data, ...)

\method{w_accuracy}{data.frame}(data, truth, estimate, na_rm = TRUE, ...)

w_accuracy_vec(truth, estimate, na_rm = TRUE, ...)
}
\arguments{
\item{data}{un \code{data.frame} que contenga \code{truth} y \code{estimate}}

\item{...}{Sin uso}

\item{truth}{columna con valor observado. Habilitada para quasiquotation}

\item{estimate}{columna con valor estimado. Habilitada para quasiquotation}

\item{na_rm}{valor \code{logical} indicando si \code{NA} deben ser omitidos}
}
\value{
el rmse normalizado por la desviacion estandar

el r2 para los percentiles extremos
}
\description{
custom_metrics

formula

nrmse vect

R squared/traditional quantiles extremos

w_accuracy
}
\concept{accuracy metrics}
\concept{class metrics}
\concept{consistency metrics}
\concept{numeric metrics}
\keyword{internal}
