% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resamples.R
\name{custom_resamples}
\alias{custom_resamples}
\alias{manual_resamples}
\alias{manual_resamples_num}
\alias{manual_resamples_class}
\alias{manual_resamples_null}
\alias{vfolds}
\alias{kfolds}
\alias{folds_vec}
\alias{invfolds_vec}
\alias{kfolds_vec}
\alias{invkfolds_vec}
\alias{sliding_resamples}
\alias{bootsplits}
\alias{random_rset}
\alias{split_resamples}
\title{helpers function}
\usage{
manual_resamples(dataset, times = 5, outcome = NULL, rango = c(0.4, 0.6))

manual_resamples_num(dataset, times = 5, outcome, rango = c(0.4, 0.6))

manual_resamples_class(dataset, times = 5, outcome, rango = c(0.4, 0.6))

manual_resamples_null(dataset, times = 5, rango = c(0.4, 0.6))

vfolds(dataset, times = 10, outcome = NULL, reverse = F)

kfolds(dataset, times = 10, outcome = NULL, rep = 1, reverse = F)

folds_vec(x, times = ncase(length(x)))

invfolds_vec(x, times = ncase(length(x)))

kfolds_vec(x, times = ncase(length(x)), rep = nsplits(length(x)))

invkfolds_vec(x, times = ncase(length(x)), rep = nsplits(length(x)))

sliding_resamples(dataset, times = 10)

bootsplits(dataset, times = 10, outcome = NULL, prop = 1)

random_rset(dataset, times = 10, rep = 3, outcome = NULL, prop = 1)

split_resamples(dataset, times = 10, split_col)
}
\description{
helpers function

Resamples
}
\examples{

data <- datalearn$iris

bootsplits(
  dataset = data,
  times = 5,
  outcome = species
)
kfolds(
  dataset = data,
  times = 5,
  outcome = species,
  rep = 3
)
manual_resamples(
  dataset = data,
  times = 5,
  outcome = species,
  rango = c(.4, .5)
)
sliding_resamples(dataset = data, times = 6)
split_resamples(
  dataset = data,
  times = 4,
  split_col = species
)
vfolds(
  dataset = data,
  times = 4,
  outcome = species
)
random_rset(
  dataset = data,
  times = 5,
  rep = 3,
  outcome = species
)

invkfolds_vec(
  x = seq(1, 1000, 1),
  times = 4,
  rep = 4
)
kfolds_vec(
  x = seq(1, 1000, 1),
  times = 4,
  rep = 4
)

}
\keyword{internal}
