% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_summary.R
\name{data_summary}
\alias{data_summary}
\alias{what_is}
\title{data_summary}
\usage{
data_summary(data)

what_is(x, ..., propr = NULL)
}
\description{
data_summary

data summary
}
\examples{

data <- mgalda::datalearn$credit
data_summary(data)
}
