% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distributions_pdqr.R
\name{dens_cumdistr_quan_rand}
\alias{dens_cumdistr_quan_rand}
\alias{dged}
\alias{pged}
\alias{qged}
\alias{rged}
\alias{dsged}
\alias{psged}
\alias{qsged}
\alias{rsged}
\alias{dsnorm}
\alias{psnorm}
\alias{qsnorm}
\alias{rsnorm}
\alias{dsstd}
\alias{psstd}
\alias{qsstd}
\alias{rsstd}
\alias{dstd}
\alias{pstd}
\alias{qstd}
\alias{rstd}
\alias{dpower}
\alias{ppower}
\alias{qpower}
\alias{rpower}
\alias{dpareto}
\alias{ppareto}
\alias{qpareto}
\alias{rpareto}
\alias{dlaplace}
\alias{plaplace}
\alias{qlaplace}
\alias{rlaplace}
\alias{dgumbel}
\alias{qgumbel}
\alias{pgumbel}
\alias{rgumbel}
\alias{drayleigh}
\alias{prayleigh}
\alias{qrayleigh}
\alias{rrayleigh}
\alias{dinvchisq}
\alias{pinvchisq}
\alias{qinvchisq}
\alias{rinvchisq}
\alias{dsc_t2}
\alias{psc_t2}
\alias{qsc_t2}
\alias{rsc_t2}
\alias{rgev}
\alias{dgev}
\alias{pgev}
\alias{qgev}
\alias{rgpd}
\alias{dgpd}
\alias{pgpd}
\alias{qgpd}
\alias{rloglap}
\alias{dloglap}
\alias{qloglap}
\alias{ploglap}
\alias{dhuber}
\alias{rhuber}
\alias{qhuber}
\alias{phuber}
\alias{dexpgeom}
\alias{pexpgeom}
\alias{qexpgeom}
\alias{rexpgeom}
\alias{dexplog}
\alias{pexplog}
\alias{qexplog}
\alias{rexplog}
\alias{dposnorm}
\alias{pposnorm}
\alias{qposnorm}
\alias{rposnorm}
\alias{dbetanorm}
\alias{pbetanorm}
\alias{qbetanorm}
\alias{rbetanorm}
\alias{dfoldnorm}
\alias{pfoldnorm}
\alias{qfoldnorm}
\alias{rfoldnorm}
\alias{dburr}
\alias{pburr}
\alias{qburr}
\alias{rburr}
\alias{dllogis}
\alias{pllogis}
\alias{qllogis}
\alias{rllogis}
\alias{dbetaprime}
\alias{pbetaprime}
\alias{qbetaprime}
\alias{rbetaprime}
\alias{dhalfnorm}
\alias{phalfnorm}
\alias{qhalfnorm}
\alias{rhalfnorm}
\alias{qchisq}
\alias{pchisq}
\alias{dchisq}
\alias{rchisq}
\alias{ddnorm}
\alias{qdnorm}
\alias{pdnorm}
\alias{rdnorm}
\alias{ddlnorm}
\alias{qdlnorm}
\alias{pdlnorm}
\alias{rdlnorm}
\alias{ddexp}
\alias{qdexp}
\alias{pdexp}
\alias{rdexp}
\alias{ddcauchy}
\alias{qdcauchy}
\alias{pdcauchy}
\alias{rdcauchy}
\alias{ddged}
\alias{qdged}
\alias{pdged}
\alias{rdged}
\alias{ddunif}
\alias{qdunif}
\alias{pdunif}
\alias{rdunif}
\alias{drndgeom}
\alias{qrndgeom}
\alias{prndgeom}
\alias{rrndgeom}
\alias{drndpois}
\alias{qrndpois}
\alias{prndpois}
\alias{rrndpois}
\alias{drndbinom}
\alias{qrndbinom}
\alias{prndbinom}
\alias{rrndbinom}
\alias{drndnbinom}
\alias{qrndnbinom}
\alias{prndnbinom}
\alias{rrndnbinom}
\alias{dzipois}
\alias{pzipois}
\alias{qzipois}
\alias{rzipois}
\title{Distributions}
\usage{
dged(x, mean = 0, sd = 1, nu = 2, log = FALSE)

pged(q, mean = 0, sd = 1, nu = 2, lower.tail = TRUE)

qged(p, mean = 0, sd = 1, nu = 2, lower.tail = TRUE)

rged(n, mean = 0, sd = 1, nu = 2)

dsged(x, mean = 0, sd = 1, nu = 2, xi = 1.5, log = FALSE)

psged(q, mean = 0, sd = 1, nu = 2, xi = 1.5)

qsged(p, mean = 0, sd = 1, nu = 2, xi = 1.5)

rsged(n, mean = 0, sd = 1, nu = 2, xi = 1.5)

dsnorm(x, mean = 0, sd = 1, xi = 1.5, log = FALSE)

psnorm(q, mean = 0, sd = 1, xi = 1.5)

qsnorm(p, mean = 0, sd = 1, xi = 1.5)

rsnorm(n, mean = 0, sd = 1, xi = 1.5)

dsstd(x, mean = 0, sd = 1, nu = 5, xi = 1.5, log = FALSE)

psstd(q, mean = 0, sd = 1, nu = 5, xi = 1.5)

qsstd(p, mean = 0, sd = 1, nu = 5, xi = 1.5)

rsstd(n, mean = 0, sd = 1, nu = 5, xi = 1.5)

dstd(x, mean = 0, sd = 1, nu = 5, log = FALSE)

pstd(q, mean = 0, sd = 1, nu = 5)

qstd(p, mean = 0, sd = 1, nu = 5)

rstd(n, mean = 0, sd = 1, nu = 5)

dpower(x, alpha, beta, log = FALSE)

ppower(q, alpha, beta, lower.tail = TRUE)

qpower(p, alpha, beta, lower.tail = TRUE)

rpower(n, alpha, beta)

dpareto(x, shape, scale, log = FALSE)

ppareto(q, shape, scale, lower.tail = TRUE, log.p = FALSE)

qpareto(p, shape, scale, lower.tail = TRUE, log.p = FALSE)

rpareto(n, shape, scale)

dlaplace(x, mu = 0, sigma = 1, log = FALSE)

plaplace(q, mu = 0, sigma = 1, lower.tail = TRUE)

qlaplace(p, mu = 0, sigma = 1, lower.tail = TRUE)

rlaplace(n, mu = 0, sigma = 1)

dgumbel(x, location = 0, scale = 1, log = FALSE)

qgumbel(p, location = 0, scale = 1, lower.tail = TRUE, log.p = FALSE)

pgumbel(q, location = 0, scale = 1, lower.tail = TRUE, log.p = FALSE)

rgumbel(n, location = 0, scale = 1)

drayleigh(x, scale = 1, log = FALSE)

prayleigh(q, scale = 1, lower.tail = TRUE, log.p = FALSE)

qrayleigh(p, scale = 1, lower.tail = TRUE, log.p = FALSE)

rrayleigh(n, scale = 1)

dinvchisq(x, df, log = FALSE)

pinvchisq(q, df, lower.tail = TRUE, log.p = FALSE)

qinvchisq(p, df, lower.tail = TRUE, log.p = FALSE)

rinvchisq(n, df)

dsc_t2(x, location = 0, scale = 1, log = FALSE)

psc_t2(q, location = 0, scale = 1, lower.tail = TRUE, log.p = FALSE)

qsc_t2(p, location = 0, scale = 1, lower.tail = TRUE, log.p = FALSE)

rsc_t2(n, location = 0, scale = 1)

rgev(n, location = 0, scale = 1, shape = 0)

dgev(x, location = 0, scale = 1, shape = 0, log = FALSE)

pgev(q, location = 0, scale = 1, shape = 0, lower.tail = TRUE, log.p = FALSE)

qgev(p, location = 0, scale = 1, shape = 0, lower.tail = TRUE, log.p = FALSE)

rgpd(n, location = 0, scale = 1, shape = 0)

dgpd(x, location = 0, scale = 1, shape = 0, log = FALSE)

pgpd(q, location = 0, scale = 1, shape = 0, lower.tail = TRUE, log.p = FALSE)

qgpd(p, location = 0, scale = 1, shape = 0, lower.tail = TRUE, log.p = FALSE)

rloglap(n, location_ald = 0, scale_ald = 1, tau = 0.5)

dloglap(x, location_ald = 0, scale_ald = 1, tau = 0.5, log = FALSE)

qloglap(
  p,
  location_ald = 0,
  scale_ald = 1,
  tau = 0.5,
  lower.tail = TRUE,
  log.p = FALSE
)

ploglap(
  q,
  location_ald = 0,
  scale_ald = 1,
  tau = 0.5,
  lower.tail = TRUE,
  log.p = FALSE
)

dhuber(x, k = 0.862, mu = 0, sigma = 1, log = FALSE)

rhuber(n, k = 0.862, mu = 0, sigma = 1)

qhuber(p, k = 0.862, mu = 0, sigma = 1, lower.tail = TRUE, log.p = FALSE)

phuber(q, k = 0.862, mu = 0, sigma = 1, lower.tail = TRUE, log.p = FALSE)

dexpgeom(x, scale = 1, shape, log = FALSE)

pexpgeom(q, scale = 1, shape)

qexpgeom(p, scale = 1, shape)

rexpgeom(n, scale = 1, shape)

dexplog(x, scale = 1, shape, log = FALSE)

pexplog(q, scale = 1, shape)

qexplog(p, scale = 1, shape)

rexplog(n, scale = 1, shape)

dposnorm(x, mean = 0, sd = 1, log = FALSE)

pposnorm(q, mean = 0, sd = 1, lower.tail = TRUE, log.p = FALSE)

qposnorm(p, mean = 0, sd = 1, lower.tail = TRUE, log.p = FALSE)

rposnorm(n, mean = 0, sd = 1)

dbetanorm(x, shape1, shape2, mean = 0, sd = 1, log = FALSE)

pbetanorm(
  q,
  shape1,
  shape2,
  mean = 0,
  sd = 1,
  lower.tail = TRUE,
  log.p = FALSE
)

qbetanorm(
  p,
  shape1,
  shape2,
  mean = 0,
  sd = 1,
  lower.tail = TRUE,
  log.p = FALSE
)

rbetanorm(n, shape1, shape2, mean = 0, sd = 1)

dfoldnorm(x, mean = 0, sd = 1, a1 = 1, a2 = 1, log = FALSE)

pfoldnorm(
  q,
  mean = 0,
  sd = 1,
  a1 = 1,
  a2 = 1,
  lower.tail = TRUE,
  log.p = FALSE
)

qfoldnorm(
  p,
  mean = 0,
  sd = 1,
  a1 = 1,
  a2 = 1,
  lower.tail = TRUE,
  log.p = FALSE
)

rfoldnorm(n, mean = 0, sd = 1, a1 = 1, a2 = 1)

dburr(x, shape1 = 2, shape2 = 1, scale = 0.5, log = FALSE)

pburr(q, shape1 = 2, shape2 = 1, scale = 0.5, log.p = FALSE, lower.tail = TRUE)

qburr(p, shape1 = 2, shape2 = 1, scale = 0.5, log.p = FALSE, lower.tail = TRUE)

rburr(n, shape1 = 2, shape2 = 1, scale = 0.5)

dllogis(x, shape, rate = 1, scale = 1/rate, log = FALSE)

pllogis(q, shape, rate = 1, scale = 1/rate, lower.tail = TRUE, log.p = FALSE)

qllogis(p, shape, rate = 1, scale = 1/rate, lower.tail = TRUE, log.p = FALSE)

rllogis(n, shape, rate = 1, scale = 1/rate)

dbetaprime(x, alpha, beta, log = FALSE)

pbetaprime(q, alpha, beta, lower.tail = TRUE, log.p = FALSE)

qbetaprime(p, alpha, beta, lower.tail = TRUE, log.p = FALSE)

rbetaprime(n, alpha, beta)

dhalfnorm(x, mean = 0, theta = sqrt(pi/2), log = FALSE)

phalfnorm(q, mean = 0, theta = sqrt(pi/2), lower.tail = TRUE, log.p = FALSE)

qhalfnorm(p, mean = 0, theta = sqrt(pi/2), lower.tail = TRUE, log.p = FALSE)

rhalfnorm(n, mean = 0, theta = sqrt(pi/2))

qchisq(p, df, ncp = 0, lower.tail = TRUE, log.p = FALSE)

pchisq(q, df, ncp = 0, lower.tail = TRUE, log.p = FALSE)

dchisq(x, df, ncp = 0, log = FALSE)

rchisq(n, df, ncp = 0)

ddnorm(x, mean, sd)

qdnorm(p, mean, sd, lower.tail = TRUE)

pdnorm(q, mean, sd, lower.tail = TRUE)

rdnorm(n, mean, sd)

ddlnorm(x, meanlog, sdlog)

qdlnorm(p, meanlog, sdlog, lower.tail = TRUE)

pdlnorm(q, meanlog, sdlog, lower.tail = TRUE)

rdlnorm(n, meanlog, sdlog)

ddexp(x, rate)

qdexp(p, rate, lower.tail = TRUE)

pdexp(q, rate, lower.tail = TRUE)

rdexp(n, rate)

ddcauchy(x, location, scale)

qdcauchy(p, location, scale, lower.tail = TRUE)

pdcauchy(q, location, scale, lower.tail = TRUE)

rdcauchy(n, location, scale)

ddged(x, mean, sd, nu)

qdged(p, mean, sd, nu, lower.tail = TRUE)

pdged(q, mean, sd, nu, lower.tail = TRUE)

rdged(n, mean, sd, nu)

ddunif(x, min, max)

qdunif(p, min, max, lower.tail = TRUE)

pdunif(q, min, max, lower.tail = TRUE)

rdunif(n, min, max)

drndgeom(x, prob, log = FALSE)

qrndgeom(p, prob, lower.tail = TRUE)

prndgeom(q, prob, lower.tail = TRUE)

rrndgeom(n, prob)

drndpois(x, lambda, log = FALSE)

qrndpois(p, lambda, lower.tail = TRUE)

prndpois(q, lambda, lower.tail = TRUE)

rrndpois(n, lambda)

drndbinom(x, size, prob, log = FALSE)

qrndbinom(p, size, prob, lower.tail = TRUE)

prndbinom(q, size, prob, lower.tail = TRUE)

rrndbinom(n, size, prob)

drndnbinom(x, size, mu, log = FALSE)

qrndnbinom(p, size, mu, lower.tail = TRUE)

prndnbinom(q, size, mu, lower.tail = TRUE)

rrndnbinom(n, size, mu)

dzipois(x, lambda, pstr0 = 0, log = FALSE)

pzipois(q, lambda, pstr0 = 0)

qzipois(p, lambda, pstr0 = 0)

rzipois(n, lambda, pstr0 = 0)
}
\description{
Distributions
}
\keyword{internal}
