% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discretizer.R
\name{discretizer}
\alias{discretizer}
\alias{discrkknn}
\alias{discrkknn.data.frame}
\alias{discrchi}
\alias{discrtopdown}
\alias{discrmdlp}
\alias{discrete}
\title{discretizer}
\usage{
discrkknn(data, ...)

\method{discrkknn}{data.frame}(data, ..., threshold = 0.1, times = 4, rep = 2, interval = c(3, 25))

discrchi(.data, y, ...)

discrtopdown(.data, y, ...)

discrmdlp(.data, y, ...)

discrete(x, ...)
}
\description{
discretizer
}
\examples{
set.seed(7)
data <- iris[, -5]
data <- dplyr::mutate_all(data, roll_noise) \%>\%
as_tbl()
object <- discrkknn(data, everything())
predict(object, data)

data <- mgalda::datalearn$credit \%>\%
  tidyr::drop_na() \%>\%
  dplyr::select(status, where(is_dblint)) \%>\%
  dplyr::group_by(status) \%>\%
  sample_n(size = 200) \%>\%
  dplyr::ungroup()

dectdiscrete(.data = data, outcome = status)

data <- datalearn$housing_market
 data <- data[,are_dblint(data)]

object <-
  discrete(x = data,outcome = sales_price)
do_discrete(x = data, outcome = sales_price)
predict(object, data)

# data <- datalearn$wa_sales_price \%>\% tidyr::drop_na()

object <-
  discrchi(.data = data, y = sales_price, where(is_dblint))
do_discrchi(.data = data, y = sales_price, where(is_dblint))
predict(object, data)

object <-
  discrmdlp(.data = data, y = sales_price, where(is_dblint))
predict(object, data)
do_discrmdlp(.data = data, y = sales_price, where(is_dblint))

object <-
  discrtopdown(.data = data, y = sales_price, where(is_dblint))
do_discrtopdown(.data = data, y = sales_price, where(is_dblint))
predict(object, data)

}
\keyword{internal}
