% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fuzzy_join.R
\name{distance_join}
\alias{distance_join}
\title{distance_join}
\usage{
distance_join(
  x,
  y,
  by = NULL,
  max_dist = 1,
  method = c("euclidean", "manhattan"),
  mode = c("inner", "left", "right", "full", "semi", "anti"),
  distance_col = NULL
)
}
\arguments{
\item{x}{A tbl}

\item{y}{A tbl}

\item{by}{Columns by which to join the two tables}

\item{max_dist}{Maximum distance to use for joining}

\item{method}{Method to use for computing distance, either euclidean (default)
or manhattan.}

\item{mode}{One of "inner", "left", "right", "full" "semi", or "anti"}

\item{distance_col}{If given, will add a column with this
name containing the distance between the two}
}
\description{
distance_join
}
\examples{

library(dplyr)

head(iris)
sepal_lengths <- data.frame(
  Sepal.Length = c(5, 6, 7),
  Sepal.Width = 1:3
)

iris \%>\%
 distance_join(sepal_lengths, max_dist = 2, mode = "inner",distance_col = "dist")
}
