% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distributions_generic_n_ops.R
\name{distr_generic_n_ops}
\alias{distr_generic_n_ops}
\alias{Math.distrobj}
\alias{Ops.distrobj}
\alias{resupport_distribution}
\alias{trasnfor_distribution}
\alias{retype_distribution}
\alias{regrid_distribution}
\alias{mix_distribution}
\alias{smooth_distribution}
\alias{estimate_distribution}
\alias{recenter_distribution}
\alias{respread_distribution}
\alias{tail_distribution}
\alias{distrstats_interval}
\alias{distrstats_center}
\alias{distrstats_metrics}
\alias{distrstats_roc}
\alias{distrstats_rocauc}
\alias{distrstats_moment}
\alias{distrstats_spread}
\alias{roc_plot}
\alias{roc_lines}
\alias{distrstats_separation}
\title{random_args}
\usage{
\method{Math}{distrobj}(x, ...)

\method{Ops}{distrobj}(e1, e2)

resupport_distribution(
  f,
  support,
  method = c("reflect", "trim", "winsor", "linear")
)

trasnfor_distribution(
  f_list,
  trans,
  ...,
  method = "random",
  n_sample = 10000,
  args_new = list()
)

retype_distribution(f, type = NULL, method = c("piecelin", "dirac", "value"))

regrid_distribution(f, n_grid, method = c("x", "q"))

mix_distribution(f_list, weights = NULL)

smooth_distribution(f, n_sample = 10000, args_new = list())

estimate_distribution(
  f,
  stat,
  sample_size,
  ...,
  n_sample = 10000,
  args_new = list()
)

recenter_distribution(f, to, method = "mean")

respread_distribution(
  f,
  to,
  method = c("var", "sd", "iqr", "mad", "range"),
  center_method = c("mean", "median", "mode", "midrange")
)

tail_distribution(f, level, method = c("trim", "winsor"), direction = "both")

distrstats_interval(
  f,
  level = 0.95,
  method = c("minwidth", "percentile", "sigma"),
  n_grid = 10001
)

distrstats_center(
  f,
  method = c("mean", "median", "mode", "midrange"),
  methods_mode = c("global", "local")
)

distrstats_metrics(f, g, threshold, method = "F1")

distrstats_roc(f, g, n_grid = 1001)

distrstats_rocauc(f, g, method = c("expected", "pessimistic", "optimistic"))

distrstats_moment(f, order)

distrstats_spread(f, method = c("var", "sd", "iqr", "mad", "range"))

roc_plot(roc, ..., add_bisector = TRUE)

roc_lines(roc, ...)

distrstats_separation(
  f,
  g,
  method = c("KS", "GM", "OP", "F1", "MCC"),
  n_grid = 10001
)
}
\description{
random_args
}
\examples{


vec_cont <- xrandom(n = 1000, mean = 2, sd = 5)
vec_cont <- safe_vec(vec_cont)
vec_dist <- rbinom(1000, 20:35, .5)

dist_cont <- density_integrate(x = vec_cont)
dist_dist <- density_integrate(x = vec_dist)

dist_cont+dist_dist
dist_cont>dist_dist
dist_cont<=dist_dist
dist_dist*2
dist_dist-2
abs(dist_cont)
sign(dist_cont)


vec_cont<-xrandom(n = 1000,mean = 20,sd = 3)
vec_cont <- safe_vec(vec_cont)
vec_dist <- rbinom(1000,20:35,.5)

dist_cont <- density_integrate(x = vec_cont)
dist_dist <- density_integrate(x = vec_dist)

resupport_distribution(f = dist_cont,support = c(18,22),method = "reflect")
retype_distribution(f = dist_cont,type = "discrete",method = "piecelin")
regrid_distribution(f = dist_cont,n_grid = 1000)
mix_distribution(f_list = list(dist_dist,dist_cont))
smooth_distribution(f = dist_cont,n_sample = 100)
estimate_distribution(f = dist_dist,stat = mean,sample_size = 100)
recenter_distribution(f = dist_cont,to = 10)
respread_distribution(f = dist_cont,to = 10)
#tail_distribution(f = dist_cont,level = .5,method = "trim")


distrstats_interval(f = dist_cont)
distrstats_center(f = dist_cont,method = "mean")
distrstats_center(f = dist_cont,method = "mode")
distrstats_moment(f = dist_cont,order = 1)
distrstats_spread(f = dist_cont,method = "var")
distrstats_spread(f = dist_cont,method = "mad")

distrstats_interval(f = dist_dist)
distrstats_center(f = dist_dist,method = "mean")
distrstats_center(f = dist_dist,method = "mode")
distrstats_moment(f = dist_dist,order = 1)
distrstats_spread(f = dist_dist,method = "var")
distrstats_spread(f = dist_dist,method = "mad")

distrstats_separation(f = dist_dist,g = dist_cont)



}
