% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distributions_obj_to_qpdr.R
\name{distr_qpdr}
\alias{distr_qpdr}
\alias{density_dist}
\alias{quantile_dist}
\title{qpdr}
\usage{
density_dist(dist, ...)

quantile_dist(dist, ...)
}
\description{
qpdr
}
\examples{

set.seed(7)
vec_cont <- rnorm(n = 1000, mean = 20, sd = 3)
vec_cont <- safe_vec(vec_cont)
vec_dist <- rbinom(1000, 20:35, .5)

## density integrate

dist_cont <- density_integrate(x = vec_cont)
dist_disc <- density_integrate(x = vec_dist)

# continua

q <- quantile_dist(dist = dist_cont, p = seq(.1, .9, .1))
q
p <- probability_dist(dist = dist_cont, q = q)
p
d <- density_dist(dist = dist_cont, x = q)
d
random_dist(dist = dist_cont, n = 10)

f_q <- as_quan(dist = dist_cont)
f_p <- as_prob(dist = dist_cont)
f_d <- as_dens(dist = dist_cont)
f_r <- as_rand(dist = dist_cont)

f_q(p = seq(.1, .9, .1)) == q
f_p(q = q) == p
f_d(x = q) == d
f_r(n = 100)

# discreta

q <- quantile_dist(dist = dist_disc, p = seq(.1, .9, .1))
q
p <- probability_dist(dist = dist_disc, q = q)
p
d <- density_dist(dist = dist_disc, x = q)
d
random_dist(dist = dist_disc, n = 10)

f_q <- as_quan(dist = dist_disc)
f_p <- as_prob(dist = dist_disc)
f_d <- as_dens(dist = dist_disc)
f_r <- as_rand(dist = dist_disc)

f_q(p = seq(.1, .9, .1)) == q
f_p(q = q) == p
f_d(x = q) == d
f_r(n = 100)

## fit dist

#rm_ls(vec_cont, vec_dist)

dist_cont <- distrcont(x = vec_cont)
dist_disc <- distrdiscr(x = vec_dist)

# continua

q <- quantile_dist(dist = dist_cont, p = seq(.1, .9, .1))
q
p <- probability_dist(dist = dist_cont, q = q)
p
d <- density_dist(dist = dist_cont, x = q)
d
random_dist(dist = dist_cont, n = 10)

f_q <- as_quan(dist = dist_cont)
f_p <- as_prob(dist = dist_cont)
f_d <- as_dens(dist = dist_cont)
f_r <- as_rand(dist = dist_cont)

f_q(p = seq(.1, .9, .1)) == q
f_p(q = q) == p
f_d(x = q) == d
f_r(n = 100)

# discreta

q <- quantile_dist(dist = dist_disc, p = seq(.1, .9, .1))
q
p <- probability_dist(dist = dist_disc, q = q)
p
d <- density_dist(dist = dist_disc, x = q)
d
random_dist(dist = dist_disc, n = 10)

f_q <- as_quan(dist = dist_disc)
f_p <- as_prob(dist = dist_disc)
f_d <- as_dens(dist = dist_disc)
f_r <- as_rand(dist = dist_disc)

f_q(p = seq(.1, .9, .1)) == q
f_p(q = q) == p
f_d(x = q) == d
f_r(n = 100)


## fit_custom_distr

#rm_ls(vec_cont, vec_dist)

dist_cont <- fit_custom_distr(x = vec_cont)
dist_disc <- fit_custom_distr(x = vec_dist)

# continua

q <- quantile_dist(dist = dist_cont, p = seq(.1, .9, .1))
q
p <- probability_dist(dist = dist_cont, q = q)
p
d <- density_dist(dist = dist_cont, x = q)
d
random_dist(dist = dist_cont, n = 10)

f_q <- as_quan(dist = dist_cont)
f_p <- as_prob(dist = dist_cont)
f_d <- as_dens(dist = dist_cont)
f_r <- as_rand(dist = dist_cont)

f_q(p = seq(.1, .9, .1)) == q
f_p(q = q) == p
f_d(x = q) == d
f_r(n = 100)

# discreta

q <- quantile_dist(dist = dist_disc, p = seq(.1, .9, .1))
q
p <- probability_dist(dist = dist_disc, q = q)
p
d <- density_dist(dist = dist_disc, x = q)
d
random_dist(dist = dist_disc, n = 10)

f_q <- as_quan(dist = dist_disc)
f_p <- as_prob(dist = dist_disc)
f_d <- as_dens(dist = dist_disc)
f_r <- as_rand(dist = dist_disc)

f_q(p = seq(.1, .9, .1)) == q
f_p(q = q) == p
f_d(x = q) == d
f_r(n = 100)


## truncate

#rm_ls(vec_cont, vec_dist,dist_cont,dist_disc)

dist_cont <-
   truncate_distr(
      dist = dist_cont,
      upper = quantile(vec_cont, .8),
      lower = quantile(vec_cont, .2)
   )
dist_disc <- truncate_distr(
   dist = dist_disc,
   upper = ceiling(quantile(vec_dist, .8)),
   lower = trunc(quantile(vec_dist, .2))
)

# continua

q <- quantile_dist(dist = dist_cont, p = seq(.1, .9, .1))
q
p <- probability_dist(dist = dist_cont, q = q)
p
d <- density_dist(dist = dist_cont, x = q)
d
random_dist(dist = dist_cont, n = 10)

f_q <- as_quan(dist = dist_cont)
f_p <- as_prob(dist = dist_cont)
f_d <- as_dens(dist = dist_cont)
f_r <- as_rand(dist = dist_cont)

f_q(p = seq(.1, .9, .1)) == q
f_p(q = q) == p
f_d(x = q) == d
f_r(n = 100)

# discreta

q <- quantile_dist(dist = dist_disc, p = seq(.1, .9, .1))
q
p <- probability_dist(dist = dist_disc, q = q)
p
d <- density_dist(dist = dist_disc, x = q)
d
random_dist(dist = dist_disc, n = 10)

f_q <- as_quan(dist = dist_disc)
f_p <- as_prob(dist = dist_disc)
f_d <- as_dens(dist = dist_disc)
f_r <- as_rand(dist = dist_disc)

f_q(p = seq(.1, .9, .1)) == q
f_p(q = q) == p
f_d(x = q) == d
f_r(n = 100)

## density

#rm_ls(vec_cont, vec_dist)
dist_cont <- density(x = vec_cont)
dist_disc <- density(x = vec_dist)

# continua

q <- quantile_dist(dist = dist_cont, p = seq(.1, .9, .1),type = "continuous")
q
p <- probability_dist(dist = dist_cont, q = q,type = "continuous")
p
d <- density_dist(dist = dist_cont, x = q,type = "continuous")
d
random_dist(dist = dist_cont, n = 10,type = "continuous")

f_q <- as_quan(dist = dist_cont,type = "continuous")
f_p <- as_prob(dist = dist_cont,type = "continuous")
f_d <- as_dens(dist = dist_cont,type = "continuous")
f_r <- as_rand(dist = dist_cont,type = "continuous")

f_q(p = seq(.1, .9, .1)) == q
f_p(q = q) == p
f_d(x = q) == d
f_r(n = 100)

# discreta

q <- quantile_dist(dist = dist_disc, p = seq(.1, .9, .1),type = "discrete")
q
p <- probability_dist(dist = dist_disc, q = q,type = "discrete")
p
d <- density_dist(dist = dist_disc, x = q,type = "discrete")
d
random_dist(dist = dist_disc, n = 10,type = "discrete")

f_q <- as_quan(dist = dist_disc,type = "discrete")
f_p <- as_prob(dist = dist_disc,type = "discrete")
f_d <- as_dens(dist = dist_disc,type = "discrete")
f_r <- as_rand(dist = dist_disc,type = "discrete")

f_q(p = seq(.1, .9, .1)) == q
f_p(q = q) == p
f_d(x = q) == d
f_r(n = 100)


}
\keyword{internal}
