% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distributions_utils.R
\name{distr_utils}
\alias{distr_utils}
\title{distr_utils}
\description{
distr_utils
}
\examples{

vec_cont<-xrandom(n = 1000,mean = 20,sd = 3)
vec_cont <- safe_vec(vec_cont)
vec_dist <- rbinom(1000,20:35,.5)

distr_cont <- distrcont(x = vec_cont)
distr_discr<- distrdiscr(x = vec_dist)

distfn_domain_num(dist = distr_cont)
distfn_domain_num(dist = distr_discr)

is_distrobj(distr_cont)
is_distrobj(distr_discr)


high_dens_int(dist = distr_cont,size = .9)
high_dens_int(dist = distr_discr,size = .9)

high_dens_vect(dist = distr_cont)
high_dens_vect(dist = distr_discr)


is_heavytail(x = distr_cont)
is_heavytail(x = distr_discr)



}
\keyword{internal}
