% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distributions.R
\name{distribution_models}
\alias{distribution_models}
\title{distrcont}
\description{
distrcont
}
\examples{

library(ggplot2)

vec_cont <- xrandom(n = 1000, mean = 20, sd = 3)
vec_cont <- safe_vec(vec_cont)
vec_dist <- rbinom(1000, 20:35, .5)

distr_cont <- distrcont(x = vec_cont)
distr_discr <- distrdiscr(x = vec_dist)

distr_cont
distr_discr

autoplot(distr_cont)
autoplot(distr_discr)

distr_cont <- fit_custom_distr(x = vec_cont)
distr_discr <- fit_custom_distr(x = vec_dist)

distr_cont
distr_discr

autoplot(distr_cont)
autoplot(distr_discr)

distr_cont <- density_integrate(x = vec_cont)
distr_discr <- density_integrate(x = vec_dist)

distr_cont
distr_discr

autoplot(distr_cont)
autoplot(distr_discr)

distr_cont <- pdqr_distr("norm", mean = 3, sd = 4)
distr_discr <- pdqr_distr(f = "rndbinom", size = 15, prob = .4)

distr_cont
distr_discr

autoplot(distr_cont)
autoplot(distr_discr)

q_cont <- quantile_dist(dist = distr_cont, p = c(.1, .9))
q_discr <- quantile_dist(dist = distr_discr, p = c(.1, .9))

distr_cont <- truncate_distr(distr_cont, q_cont[2], q_cont[1])
distr_discr <- truncate_distr(distr_discr, q_discr[2], q_discr[1])

distr_cont
distr_discr

autoplot(distr_cont)
autoplot(distr_discr)


distr_cont <- fit_pdqr(x = vec_cont, fn = "norm")
distr_discr <- fit_pdqr(x = vec_dist, fn = "rndbinom")

distr_cont
distr_discr

autoplot(distr_cont)
autoplot(distr_discr)
}
\keyword{internal}
