% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statisticstools.R
\name{efron_rsq}
\alias{efron_rsq}
\title{Efron's Pseudo R-sqaured}
\usage{
efron_rsq(x, lf)
}
\arguments{
\item{x}{a data matrix.}

\item{lf}{Observed logistic factors.}
}
\description{
Efron's Pseudo R-sqaured
}
\author{
Wei Hao
}
