% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f_ply.R
\name{f_ply}
\alias{f_ply}
\alias{lgl_all}
\alias{lgl_any}
\alias{chr_cllp}
\alias{cmp_map}
\alias{cmp_map2}
\alias{cmp_pmap}
\alias{map_maxdepth}
\alias{flatten}
\alias{cmp_flatten}
\alias{cmp_ftn_map}
\alias{cmp_ftn_map_if}
\alias{vaggr}
\alias{vaggr_mean}
\alias{slide_dbl}
\alias{simplify}
\alias{.un}
\alias{probe}
\alias{map2_if}
\alias{imap_if}
\title{Apply func}
\usage{
lgl_all(.x, .f, ...)

lgl_any(.x, .f, ...)

chr_cllp(.x, .f, ..., .collapse = "")

cmp_map(.x, .f, ...)

cmp_map2(.x, .y, .f, ...)

cmp_pmap(.l, .f, ...)

map_maxdepth(.x, .f, ...)

flatten(.x)

cmp_flatten(.x)

cmp_ftn_map(.x, .f, ...)

cmp_ftn_map_if(.x, .p, .f, ..., .else = NULL)

vaggr(x, .group, .fn, ..., .default = NULL)

vaggr_mean(x, .group)

slide_dbl(
  .x,
  .fn,
  ...,
  .size = 1,
  .partial = c("none", "left", "right", "center"),
  .fill = TRUE
)

simplify(.x, .type = NULL)

.un(x)

probe(.x, .p, ...)

map2_if(.x, .y, .p, .f, ...)

imap_if(.x, .p, .f, ...)
}
\arguments{
\item{...}{other arguments passed on to \code{.fun}}

\item{x}{vector of values to aggregate}

\item{.group}{grouping vector}

\item{.fn}{aggregation function}

\item{.default}{default value used for missing groups.  This argument is
also used as the template for function output.}
}
\description{
Apply func
}
\examples{
# Some examples of use borrowed from ?tapply
n <- 17; fac <- factor(rep(1:3, length.out = n), levels = 1:5)
table(fac)
vaggr(1:n, fac, sum)
vaggr(1:n, fac, sum, .default = NA_integer_)
vaggr(1:n, fac, range)
vaggr(1:n, fac, range, .default = c(NA, NA) + 0)
vaggr(1:n, fac, quantile)

# But it is about 10x faster
x <- rnorm(1e6)
y1 <- sample.int(10, 1e6, replace = TRUE)
system.time(tapply(x, y1, mean))
system.time(vaggr(x, y1, mean))
}
\keyword{internal}
