% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/financialmath.R
\name{financialmath}
\alias{financialmath}
\alias{discounted_cf}
\alias{disc_payback}
\alias{irr}
\alias{netpresentvalue}
\alias{payback}
\alias{annuity_value}
\alias{rstated2annual}
\alias{err}
\alias{profitability_index}
\title{financialmath}
\usage{
discounted_cf(cf, r, date = NULL)

disc_payback(cf, r, date = NULL)

irr(cf, date = NULL)

netpresentvalue(cf, r, date = NULL)

payback(cf, date = NULL)

annuity_value(cf, r, date = NULL)

rstated2annual(r, m)

err(cf, rate, date = NULL)

profitability_index(cf, capex, MARR, date = NULL)
}
\description{
financialmath
}
\examples{
library(lubridate)
set.seed(7)

cf <- c(-2000,runif(20,400,700))
dates <- today() +years(0:20)
r <- .12

annuity_value(cf =cf,r = r,date = dates)
disc_payback(cf = cf,r = r,date = dates)
discounted_cf(cf = cf,r = r,date = dates)
err(cf = cf,rate = .03,date = dates)
irr(cf = cf,date = dates)
netpresentvalue(cf = cf,r = r,date = dates)
payback(cf = cf,date = dates)
profitability_index(cf = cf[-1],capex = cf[1],MARR = r,date = dates[-1])
rstated2annual(r = r,m = 12)

}
\keyword{internal}
