% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_globals.R
\name{find_globals}
\alias{find_globals}
\alias{make_code_walker}
\alias{walk_code}
\title{find_globals}
\usage{
make_code_walker(
  ...,
  handler = function(v, w) {
     NULL
 },
  call = function(e, w) {
     for (ee in as.list(e)) {
         if (!missing(ee)) {
  
              walk_code(ee, w)
         }
     }
 },
  leaf = function(e, w) {
     print(e)
 }
)

walk_code(e, w = make_code_walker())

find_globals(fun, merge = TRUE)
}
\description{
find_globals

find_globals

find_globals
}
