% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphicsR_scales_palettes.R
\name{grf_xscale}
\alias{grf_xscale}
\title{grf_scale}
\usage{
grf_xscale(object, .scale, .format = FALSE)
}
\arguments{
\item{.scale}{axis scale. Allowed values are
one of c("none", "log2", "log10",
"sqrt", "percent", "dollar", "scientific",
"numeric,"millones,"miles); e.g.: .scale="log2".}

\item{.format}{ogical value. If TRUE, axis tick
mark labels will be formatted
when .scale  = "log2" or "log10".}
}
\description{
grf_scale
}
