% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mics.R, R/utils.R
\name{helpers_function}
\alias{helpers_function}
\alias{create_testthatfiles}
\alias{get_predictorclass}
\alias{check_numeric}
\alias{check_na}
\alias{arg_cols_exist}
\alias{check_rango_muestras}
\alias{names0}
\alias{vec2table}
\alias{update_namespace}
\alias{calc_dots}
\alias{ndim2point}
\alias{is_df}
\alias{is_numinvalid}
\alias{is_even}
\alias{is_fctchr}
\alias{is_dblint}
\alias{is_odd}
\alias{is_empty}
\alias{is_true}
\alias{is_false}
\alias{is_valid}
\alias{\%l0\%}
\alias{\%||\%}
\alias{\%@\%}
\alias{\%A\%}
\alias{\%nin\%}
\alias{\%notin\%}
\alias{\%out\%}
\alias{\%include\%}
\alias{\%inside\%}
\alias{random_vec}
\alias{libs_packs}
\alias{random_round}
\alias{format_ch_vec}
\title{helpers function}
\usage{
create_testthatfiles()

get_predictorclass(.data, outcome)

check_numeric(.data)

check_na(dataset, x)

arg_cols_exist(dataset, ...)

check_rango_muestras(rango)

names0(x, prefix = "id")

vec2table(truth, estimate, dnn = c("Prediction", "Truth"), ...)

update_namespace()

calc_dots(sf_data, col_names, n_per_dot)

ndim2point(x, y)

is_df(x)

is_numinvalid(x)

is_even(x)

is_fctchr(x)

is_dblint(x)

is_odd(x)

is_empty(x)

is_true(x)

is_false(x)

is_valid(x)

x \%l0\% y

x \%||\% y

x \%@\% y

x \%A\% y

x \%nin\% vect

x \%notin\% y

x \%out\% y

x \%include\% y

x \%inside\% y

random_vec(x, rango = c(0.25, 0.75))

libs_packs()

random_round(x)

format_ch_vec(x, sep = ", ", width = options()$width - 9)
}
\description{
helpers function

helpers function

Helpers for printing step functions
}
\keyword{internal}
