% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mics.R, R/utils.R
\docType{data}
\name{hfrec_libs}
\alias{hfrec_libs}
\alias{lm_libs}
\alias{hfrec_libraries}
\title{librerias de alta frecuencia}
\format{
An object of class \code{character} of length 29.

An object of class \code{character} of length 20.
}
\usage{
hfrec_libs

lm_libs

hfrec_libraries(digitos_decimales = 2, scipen = 3, mlearning_libs = F, log = F)
}
\description{
librerias de alta frecuencia
}
\keyword{internal}
