% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster_tools.R
\name{hierarchical}
\alias{hierarchical}
\title{Hierachical clustering
Convenient methods for hierachical clustering}
\usage{
hierarchical(x, method = "complete", metric = "euclidean", n = 5)
}
\arguments{
\item{x}{data frame}

\item{method}{method to use, see \code{\link{hclust}}}

\item{metric}{distance metric to use, see \code{\link{dist}}}

\item{n}{number of clusters to retrieve, see \code{\link{cut}}}
}
\description{
Hierachical clustering
Convenient methods for hierachical clustering
}
