% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster_tools.R
\name{hkmeans}
\alias{hkmeans}
\title{Hierarchical k-means clustering}
\usage{
hkmeans(
  x,
  k,
  hc.metric = "euclidean",
  hc.method = "ward.D2",
  iter.max = 10,
  km.algorithm = "Hartigan-Wong"
)
}
\description{
Hierarchical k-means clustering
}
\examples{
# Scale the data
df <- scale(iris[,1:4])

# Compute hierarchical k-means clustering
res.hk <- hkmeans(df, 4)

# Elements returned by hkmeans()
names(res.hk)

# Print the results
res.hk
}
