% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster_tools.R
\name{knn_classify}
\alias{knn_classify}
\title{with classifly.}
\usage{
knn_classify(data, ...)
}
\arguments{
\item{data}{training data set}

\item{formula}{classification formula}

\item{k}{number of neighbours to use}
}
\description{
with classifly.
}
\keyword{classif}
