% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/limits_bounds.R
\name{limits_bounds}
\alias{limits_bounds}
\alias{limits_bounds.default}
\alias{limits_object}
\alias{predict.limits_bounds}
\alias{step_limits_bounds}
\title{limites}
\usage{
limits_bounds(.data, ...)

\method{limits_bounds}{default}(.data, ...)

limits_object(x)

\method{predict}{limits_bounds}(object, new_data = NULL)

step_limits_bounds(
  recipe,
  ...,
  role = NA,
  trained = FALSE,
  object = NULL,
  skip = FALSE,
  id = rand_id("limits_bounds")
)
}
\value{
objeto clase limits_bounds
}
\description{
limites

limites

predict bounds
}
\concept{preprocessing}
\keyword{datagen}
\keyword{internal}
