% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mgalda-package.R
\docType{package}
\name{mgalda-package}
\alias{mgalda}
\alias{mgalda-package}
\title{mgalda: 360 Multi tool library}
\description{
Libreria con el fin de aprendizaje (compac) y generar herramientas.
}
\author{
\strong{Maintainer}: Matias G. \email{mgaldame@rfen.uchile.cl}

}
\keyword{internal}
