% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mics.R
\docType{data}
\name{base_class_otras}
\alias{base_class_otras}
\alias{base_class}
\alias{base_reg}
\alias{curvas_df}
\alias{datalearn}
\alias{geocodigos}
\alias{list_sample}
\title{pkg_data}
\format{
A data frame with 3 variables

A data frame with 3 variables

A data frame with 3 variables

dataframe con ejemplos de eerr de retail

dataframe ames reducida

dataframe ames reducida

dataframe ames reducida
}
\source{
elavoracion propia

modeldata

modeldata

modeldata
}
\usage{
base_class_otras

base_class

base_reg

curvas_df

datalearn

geocodigos

list_sample
}
\description{
pkg_data
}
\keyword{datasets}
