% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statisticstools.R
\name{pseudo_rsq}
\alias{pseudo_rsq}
\title{Mcfadden's Pseudo R-sqaured}
\usage{
pseudo_rsq(x, lf_alt, lf_null = NULL)
}
\arguments{
\item{x}{a data matrix.}

\item{lf_alt}{Observed logistic factocbind_wrapsrs.}

\item{lf_null}{Null logistic factors.}
}
\description{
Mcfadden's Pseudo R-sqaured
}
