% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_base.R, R/graphicsR_utils.R,
%   R/graphicsR.R, R/treetools.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{.data}
\alias{.env}
\alias{:=}
\alias{ggproto}
\alias{fortify}
\alias{ggplot_add}
\alias{autoplot}
\alias{as.phylo}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{ape}{\code{\link[ape]{as.phylo}}}

  \item{ggplot2}{\code{\link[ggplot2]{autoplot}}, \code{\link[ggplot2]{fortify}}, \code{\link[ggplot2]{ggplot_add}}, \code{\link[ggplot2]{ggproto}}}

  \item{rlang}{\code{\link[rlang:dot-data]{.data}}, \code{\link[rlang:dot-data]{.env}}, \code{\link[rlang:dyn-dots]{:=}}}
}}

