% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seqs.R
\name{seqs}
\alias{seqs}
\alias{seq2}
\alias{seq2_along}
\alias{seq_row}
\alias{seq_col}
\alias{seq_between}
\alias{seq_power}
\alias{seq_log}
\alias{seq_range}
\alias{seq_full}
\alias{rep_each_n}
\alias{rep_full}
\alias{rep_each}
\alias{rep_along}
\title{Increasing sequence of integers in an interval}
\usage{
seq2(from, to)

seq2_along(from, x)

seq_row(x)

seq_col(x)

seq_between(range, ...)

seq_power(from, to, length.out, degree = 2)

seq_log(
  from,
  to,
  length.out,
  base = 10,
  plus1 = TRUE,
  start = c("lower", "mid", "upper")
)

seq_range(x, n, by, trim = NULL, expand = NULL, pretty = FALSE)

seq_full(range, size, pad = FALSE)

rep_each_n(n, times)

rep_full(x, times, length)

rep_each(x, times)

rep_along(x, along)
}
\arguments{
\item{from}{The starting point of the sequence.}

\item{to}{The end point.}

\item{x}{A vector whose length is the end point in \code{seq2_along()} or
a data frame \code{seq_col()} or \code{seq2_row()}}

\item{range}{range de seq}
}
\value{
An integer vector containing a strictly increasing
sequence.
}
\description{
Increasing sequence of integers in an interval
}
\examples{
seq2(2, 10)
seq2(10, 2)
seq(10, 2)

seq2_along(10, letters)

seq_row(mtcars)

seq_col(mtcars)

seq_between(range = c(4, 10))

seq_log(from = 1, to = 1000, length.out = 10, base = 10)

seq_power(from = 1, to = 1000, length.out = 10, degree = 3)

x <- rcauchy(100)
seq_range(x, n = 10)
seq_range(x, n = 10, trim = 0.1)
seq_range(x, by = 1, trim = 0.1)

# Make pretty sequences
y <- runif(100)
seq_range(y, n = 10)
seq_range(y, n = 10, pretty = TRUE)
seq_range(y, n = 10, expand = 0.5, pretty = TRUE)

seq_range(y, by = 0.1)
seq_range(y, by = 0.1, pretty = TRUE)

}
