% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mltools.R
\name{smooth_linear}
\alias{smooth_linear}
\alias{smooth_linear.default}
\alias{smooth_linear.data.frame}
\alias{smooth_linear.matrix}
\alias{smooth_linear.formula}
\alias{smooth_linear.recipe}
\title{Fit a smooth_linear}
\usage{
smooth_linear(x, ...)

\method{smooth_linear}{default}(x, ...)

\method{smooth_linear}{data.frame}(x, y, times = 5, rep = 3)

\method{smooth_linear}{matrix}(x, y, times = 5, rep = 3)

\method{smooth_linear}{formula}(formula, data, times = 5, rep = 3)

\method{smooth_linear}{recipe}(x, data, times = 5, rep = 3)
}
\description{
Fit a smooth_linear
}
