% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/superlearner_utils.R
\name{superlearner_utils}
\alias{superlearner_utils}
\alias{outcome_expr}
\alias{outcome_quo}
\alias{outcome_chr}
\title{superlearner_utils}
\usage{
outcome_expr(x)

outcome_quo(x)

outcome_chr(x)
}
\description{
superlearner_utils
}
\keyword{internal}
