% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphicsR.R
\name{theme_ggplot}
\alias{theme_ggplot}
\alias{theme_mgg}
\alias{theme_agile}
\alias{theme_sr}
\alias{theme_plex}
\alias{theme_dotdens}
\alias{theme_roboto}
\title{herramientas ggplot}
\usage{
theme_agile(
  base_size = 11,
  base_family = "Arial",
  lines_lwd = 0.5,
  plot_grid = TRUE,
  base_col = "black",
  axis_lines = TRUE,
  ticks_type = "outer",
  alpha_leg = 0.1,
  bord_size = 0,
  background_colour = "white",
  grid_thick = 1,
  grid_type = "solid",
  ticks_xy = "xy",
  grid_cols = c("grey50", "grey70")
)

theme_mgg(...)

theme_sr(...)

theme_plex(
  base_size = 11,
  strip_text_size = 12,
  strip_margin = 5,
  subtitle_margin = 10,
  title_margin = 10,
  ...
)

theme_dotdens(
  legend = TRUE,
  legend_size = 10,
  background_colour = "#212121",
  base_col = "white"
)

theme_roboto(
  base_size = 11,
  strip_text_size = 12,
  strip_margin = 5,
  subtitle_margin = 10,
  title_margin = 10,
  ...
)
}
\description{
herramientas ggplot

ggplot theme

ggplot theme

ggplot theme

Minimal ggplot2 IBM Plex Sans fonts

dens

Minimal ggplot2 roboto
}
\keyword{internal}
