% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/twostep.R
\name{twostep}
\alias{twostep}
\title{modelo en dos pasos}
\usage{
twostep(first_step, second_step, first_class)
}
\arguments{
\item{first_step}{df}

\item{second_step}{submodelos}

\item{first_class}{numero de resamples}
}
\value{
modelo de clasificacion en dos pasos
}
\description{
modelo en dos pasos
}
