% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_base.R
\docType{data}
\name{util_base}
\alias{util_base}
\alias{base_superlearner}
\alias{curv_data}
\alias{datalearn}
\alias{geocodigos}
\alias{list_sample}
\alias{london_gpkg}
\alias{yardsticks_metrics_info}
\alias{info_distribucion}
\alias{stocks_data}
\alias{streets}
\alias{map_hex}
\alias{map_grid}
\alias{map_pois}
\alias{map_fromto}
\alias{hexbin_data}
\alias{rm_ls}
\alias{hfrec_libraries}
\title{util_base}
\format{
data

data

data

data

data

data

data

data

data

data

data

data

data

data

data
}
\usage{
base_superlearner

curv_data

datalearn

geocodigos

list_sample

london_gpkg

yardsticks_metrics_info

info_distribucion

stocks_data

streets

map_hex

map_grid

map_pois

map_fromto

hexbin_data

rm_ls(...)

hfrec_libraries(
  conflict_prefer = getOption("mgalda.conflict_prefer", TRUE),
  hfrec_libs = getOption("mgalda.hfrec_libs"),
  verbose = TRUE
)
}
\description{
util_base
}
\keyword{datasets}
\keyword{internal}
