% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_print.R
\name{utils_print}
\alias{utils_print}
\alias{cat_missing}
\alias{cat_stop}
\alias{cat_warn}
\alias{cat_message}
\alias{cat_return_null}
\alias{cat_return_false}
\alias{cat_line}
\alias{cat_step}
\alias{cat_cumstyle}
\alias{cat_title_head}
\alias{cat_endline}
\alias{cat_subtitle1}
\alias{cat_subtitle2}
\alias{color_prompt}
\alias{cat_prompt}
\alias{generic_print}
\alias{subgeneric_print}
\alias{generic_default}
\alias{osize}
\alias{recipes_boilerplate}
\alias{paste_test_pgk}
\alias{paste_examples_pgk}
\title{utils_print}
\usage{
cat_missing(var_name, value_name)

cat_stop(..., .sep = "", .envir = parent.frame(), trace = FALSE)

cat_warn(..., .sep = "", .envir = parent.frame())

cat_message(
  ...,
  .sep = "",
  type = c("success", "info", "go"),
  .envir = parent.frame()
)

cat_return_null(
  ...,
  msg = NULL,
  lgl_sum = c("any", "all"),
  type = c("warning", "stop", "info"),
  .envir = parent.frame()
)

cat_return_false(
  ...,
  msg = NULL,
  lgl_sum = c("any", "all"),
  type = c("warning", "stop", "info"),
  .envir = parent.frame()
)

cat_line(..., file = stdout())

cat_step(
  mgs,
  head = "i",
  head_sep = "-",
  .envir = env_prnt(),
  width.cutoff = getOption("width", 70),
  msg_type = "grey",
  head_type = "blue"
)

cat_cumstyle(text, background = NULL, style = NULL, color = NULL)

cat_title_head(
  text,
  .sep = "",
  line_color = "silver",
  text_color = "blue",
  .envir = parent.frame(),
  width = getOption("width", 80)
)

cat_endline(line_color = "silver", width = getOption("width", 80))

cat_subtitle1(
  text,
  sub_text = NULL,
  .sep = "",
  color = "none",
  .envir = parent.frame(),
  width = min(c(getOption("width", 30), 30))
)

cat_subtitle2(
  text,
  sub_text = NULL,
  .sep = "",
  line_color = "none",
  text_color = "none",
  subtext_color = "silver",
  .envir = parent.frame(),
  width = getOption("width", 80)
)

color_prompt(prompt, type, obj = "color")

cat_prompt(..., .sep = "", .envir = parent.frame(), type, .msg_fn)

generic_print(x, .add_print = NULL)

subgeneric_print(x, nm)

generic_default(call = sys.call(sys.parent()), envir = parent.frame(2L))

osize(x, ...)

recipes_boilerplate(
  name,
  which = c("step", "check"),
  .o = character(0),
  .init = character(0)
)

paste_test_pgk(x)

paste_examples_pgk(x)
}
\description{
utils_print
}
\keyword{internal}
