% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mltools.R
\name{xlm}
\alias{xlm}
\alias{xlm.default}
\alias{xlm.data.frame}
\alias{xlm.matrix}
\alias{xlm.formula}
\alias{xlm.recipe}
\title{Fit a xlm}
\usage{
xlm(x, ...)

\method{xlm}{default}(x, ...)

\method{xlm}{data.frame}(x, y, times = 5, rep = 3)

\method{xlm}{matrix}(x, y, times = 5, rep = 3)

\method{xlm}{formula}(formula, data, times = 5, rep = 3)

\method{xlm}{recipe}(x, data, times = 5, rep = 3)
}
\description{
Fit a xlm
}
