% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mltools.R
\name{xloess}
\alias{xloess}
\alias{xloess.default}
\alias{xloess.data.frame}
\alias{xloess.matrix}
\alias{xloess.formula}
\alias{xloess.recipe}
\title{Fit a xloess}
\usage{
xloess(x, ...)

\method{xloess}{default}(x, ...)

\method{xloess}{data.frame}(x, y, times = 5, rep = 3)

\method{xloess}{matrix}(x, y, times = 5, rep = 3)

\method{xloess}{formula}(formula, data, times = 5, rep = 3)

\method{xloess}{recipe}(x, data, times = 5, rep = 3)
}
\description{
Fit a xloess
}
