% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xnls.R
\name{xnls}
\alias{xnls}
\alias{xnls.default}
\alias{step_xnls}
\title{xnls}
\usage{
xnls(.data, x, y)

\method{xnls}{default}(.data, x, y)

step_xnls(
  recipe,
  ...,
  role = NA,
  trained = FALSE,
  object = NULL,
  skip = FALSE,
  id = rand_id("xnls")
)
}
\value{
objeto clase xnls
}
\description{
xnls

xnls
}
\concept{preprocessing}
\keyword{datagen}
\keyword{internal}
