% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/principal_component.R
\name{xpca}
\alias{xpca}
\alias{xpca.default}
\alias{predict.xpca}
\alias{pca_coefs}
\alias{pca_variances}
\alias{pca_var}
\alias{print.xpca}
\title{Principal Components Analysis}
\usage{
xpca(.data, ..., num_comp = 3, threshold = 0.9)

\method{xpca}{default}(.data, ..., num_comp = 3, threshold = 0.9)

\method{predict}{xpca}(object, new_data)

pca_coefs(x)

pca_variances(x)

pca_var(dataset, scale = TRUE, center = TRUE)

\method{print}{xpca}(x, print.x = FALSE, ...)
}
\value{
pca
}
\description{
Principal Components Analysis

transformar pca

predict xpca
}
\keyword{internal}
