% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xrossdummy.R
\name{xrossdummy}
\alias{xrossdummy}
\alias{xrossdummy.default}
\alias{predict.xrossdummy}
\alias{step_xrossdummy}
\alias{print.xrossdummy}
\title{xrossdummy}
\usage{
xrossdummy(.data, dummy_var, ...)

\method{xrossdummy}{default}(.data, dummy_var, ...)

\method{predict}{xrossdummy}(object, new_data)

step_xrossdummy(
  recipe,
  ...,
  role = NA,
  trained = FALSE,
  dummy_var = NA,
  object = NULL,
  skip = FALSE,
  id = recipes::rand_id("xrossdummy")
)

\method{print}{xrossdummy}(x)
}
\value{
crossdummy
}
\description{
xrossdummy

crossdummy

predict sl xrossdummy

xrossdummy step
}
\concept{preprocessing}
\keyword{datagen}
\keyword{internal}
